Refactoring and cleaning up
